+++ CHANGELOG.md
@@ -26,6 +26,11 @@ v4.2
 - Fix OpenSense calibration algorithm to handle models facing an arbitrary direction. The calibration algorithm now aligns one axis of the provided Orientation Sensor data with the x-axis of the base segment (e.g. pelvis) of the model in default pose.
 - For PrescribedController, the controls_file column labels can now be absolute paths to actuators (previously, the column labels were required to be actuator names).
 - Fixed a critical bug in Induced Accelerations Analysis which prevents analysis to run when external forces are present ([PR #2847](https://github.com/opensim-org/opensim-core/pull/2808)).
+- Removed `Path` abstract base class (PR #2844)
+  - Unused by OpenSim and related projects
+- Improved the performance of `ComponentPath` (PR #2844)
+  - This improves the performance of component-heavy models by ~5-10 %
+  - The behavior and interface of `ComponentPath` should remain the same
 - The new Matlab CustomStaticOptimization.m guides the user to build their own custom static optimization code.